backend/requirements.txt
flask
flask-cors
sentence-transformers
numpy<2.0
groq
python-dotenv
